fix: change approach to trigger-image-deploy (see body)
Now, 'trigger-image-deploy' action uses the 'yq' package to set the
specified Docker image tag in an existing, non-versioned docker-compose
file, which is then passed down to the 'docker-compose' command with the
'-f' option.
For this to work, "docker-compose-override-file-name" must point to an
existing file in thedigbee/devops with the following default structure:
```yml
version: '3.7'
services:
  digbee-core:
    environment:
      NO_OP: ''
  digbee-front-server:
    environment:
      NO_OP: ''
  digbee-pdf-view-generator:
    environment:
      NO_OP: ''
  digbee-pdf-renderer:
    environment:
      NO_OP: ''
```
We must be vigilant when creating/removing services so as not to forget
to keep this file up-to-date.
Ideally, this file won't be under version control since it'll be relevant
only to the remote environment.